GH-4 Working on correcting issues in the unit tests
